requirements-lock.txt
absl-py==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.6
filelock==3.25.2
fsspec==2026.2.0
grpcio==1.78.0
idna==3.11
Jinja2==3.1.6
Markdown==3.10.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
simsimd==6.5.16
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
